backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
httpx
python-dotenv
APScheduler
anthropic
